Interoperability + simplified README
Added interoperability with blocking code.
Added interoperability namespace for Java 8 CompletableFuture.
Documented interoperability with Manifold, promesa and core.async.
Removed the handling of all exceptions thrown in callbacks. This is now responsibility of interop code.
Documented that user is expected not to throw in their callbacks.
Reduced the size of README using more concise wording and moving some topics to docstrings.